feature: Add DebugConfigurationProvider
Adds a DebugConfigurationProvider. This allows to debug any executable
configured in the project via the F5 shortcut.

vscode is a bit weird here. To pick the correct debug command it's
necessary to pick the general C/C++ debug and then the choice appears.
Once a dynamic configuration is selected, it's a toplevel choice and F5
will just rerun the last one.