Activate the extension when meson.build file is anywhere in the workspace.
This fixes logic mismatch between the activation events and the check for
existence of project inside the workspace. When meson file is outside
the root directory, the extension will still activate when the you open
the file.

The extension is still not able to handle non-root meson file, but at
least now the failure mode is promptly exposed, instead of being hidden.