Refactor support for multiple Meson projects
The very first thing the extension must do is determining the Meson
project source directory. If there are multiple workspaces, or if the
repository contains multiple projects in subdirs, there could be more
than one potential Meson project. In that case ask the user to select
one. The build directory can then be determined relative to the selected
source directory.

This also adds a command to allow switching to a different Meson project
root directory. Unfortunately this requires a full window reload for
now.