Make sure rust-analyser uses rust-project.json Meson generates
This is very similar case than compile_commands.json, Meson generates a
file in builddir that some extensions should be using. It makes static
analysers work out of the box instead of requiring user to configure
paths themself.

We only need to watch for file creation, it was previously wrongly
watching for file changes for legacy reasons (we used to patch
compile_commands.json).