utils: Add extra environment argument to exec()
It's already possible to add environment variables in the 'options'
argument — but child_process.execFile() does not treat options.env as
extra environment variables to add. It replaces the environment with
options.env entirely. So, it's convenient to have an extraEnv argument
with variables to add.

This will be necessary for the testEnvironment setting added in the next
commit.