Add custom configuration provider for cpptools
Untill now the meson extension provided C++ linting through
compile_commands.json. This would not configure:
 - New file that haven't been compiled yet
 - External libraries that are not part of the project (when browsing/debugging)
This was especially problematic when using a cross-compilation toolchain
since none of the default host headers should be used.

This adds a custom configuration provider for the cpptools extension. It
introspects the compiler settings of the project to configure linting
and browsing. Currently, file-specific configurations are not supported.
Some other advanced meson options might also not be supported. However
it provides a strong experience for cross-compilation and external
library support.

Here is an example settings.json configuration to activate this provider:
```json
{
  "C_Cpp.default.configurationProvider": "mesonbuild.mesonbuild"
}
```