Categorise test cases into test suite(s)
Meson test cases may belong to test suite(s). If so, in VSCode's
"Testing" feature from the Activity Bar, add the test suite(s) as a
parent item and add the test cases as children.

The user has the choice to run all tests belonging to a test suite.
Will pass the test suite to the meson test command --suite. The option
--suite is needed to run test cases that the meson config purposely
excludes (for example, test suite "flaky").

Fixes #245

Quote the "test case" in the meson command, so can run test cases with a
space in the name. The quotes are needed because `cp.ExecFileOptions`
sets `shell=true`.

Fixes #256