update (almost) all dependencies
In particular, bump some outdated dev dependencies
I've removed markdownlint because it's responsible for > 50% of
yarn.lock and a transitive dependency requires DOM types, which in turn
are not available in nodejs

Also bump to vscode 1.101 for newer nodejs and ESM support.
This version is 6 months old at the time of writing.

Signed-off-by: Jannik Glückert <jannik.glueckert@gmail.com>